Refactor HarpFeatures to be in inventory
